fix: simplify app tsx
